Move CopyEliminationAnalyzer to 'copy' package
