+++ modules/wyopcl/src/wyopcl/translator/Translator.java
@@ -24,6 +24,7 @@ import wyil.lang.WyilFile;
 import wyil.lang.WyilFile.FunctionOrMethod;
 import wyil.util.AttributedCodeBlock;
 import wyopcl.Configuration;
+import wyopcl.translator.copy.CopyEliminationAnalyzer;
 import wyopcl.translator.generator.CodeGenerator;
 import wyopcl.translator.generator.CodeGeneratorHelper;
 import wyopcl.translator.symbolic.PatternMatcher;
